requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
